Use stringTemplateMember
